Structured around 95% of each route.
Most routes have a completed design but a few still need grid allocation and API implementation. Created another JS file and began the finance tool portion of the project. Also, added 2 new NPM packages, which are the "us-state-codes" and "zipcodes" packages that made it easier for me to get around the API calls accurately.